CON-167: Move BlockStoreBenchSuite to separate file
