+++ block-storage/src/test/scala/io/casperlabs/blockstorage/benchmarks/BlockStoreBench.scala
@@ -3,28 +3,13 @@ package io.casperlabs.blockstorage.benchmarks
 import java.nio.file.Paths
 
 import cats.Monad
-import cats.syntax.traverse._
 import cats.instances.list._
-import com.google.protobuf.ByteString
-import io.casperlabs.blockstorage.BlockStore.BlockHash
-import io.casperlabs.blockstorage.{
-  BlockStore,
-  FileLMDBIndexBlockStore,
-  InMemBlockStore,
-  LMDBBlockStore
-}
-import io.casperlabs.casper.protocol.{BlockMessage, Header, Justification}
-import io.casperlabs.{metrics, shared}
-import io.casperlabs.metrics.Metrics
-import io.casperlabs.storage.BlockMsgWithTransform
+import cats.syntax.traverse._
+import BlockStoreBenchSuite._
+import io.casperlabs.blockstorage._
 import monix.eval.Task
 import monix.execution.Scheduler.Implicits.global
 import org.openjdk.jmh.annotations._
-
-import scala.collection.immutable.IndexedSeq
-import scala.util.Random
-import io.casperlabs.blockstorage.benchmarks.BlockStoreBenchSuite._
-import io.casperlabs.shared.Log
 
 @State(Scope.Benchmark)
 @BenchmarkMode(Array(Mode.Throughput))
@@ -106,68 +91,4 @@ class FileLMDBIndexBench extends BlockStoreBench {
       .runSyncUnsafe()
       .right
       .get
-}
-
-object BlockStoreBenchSuite {
-  import scala.language.implicitConversions
-
-  implicit def strToByteStr(str: String): ByteString =
-    ByteString.copyFromUtf8(str)
-
-  implicit val metricsNop: Metrics[Task] = new metrics.Metrics.MetricsNOP[Task]
-  implicit val logNop: Log[Task]         = new shared.Log.NOPLog[Task]
-
-  private val preCreatedBlocks = (0 to 50000).map(_ => randomBlock)
-
-  val blocksIter = repeatedIteratorFrom(preCreatedBlocks)
-  val preFilling = (0 to 10000).map(_ => randomBlock)
-
-  def randomInserted =
-    preFilling(Random.nextInt(preFilling.size))
-
-  def randomHexString(numchars: Int): String = {
-    val sb = new StringBuffer
-    while (sb.length < numchars) sb.append(Integer.toHexString(Random.nextInt))
-
-    sb.toString.substring(0, numchars)
-  }
-
-  def randomBlock: (BlockHash, BlockMsgWithTransform) =
-    (randomHash, randomBlockMessage)
-
-  def randomHash: BlockHash =
-    strToByteStr(randomHexString(32))
-
-  def randomBlockMessage: BlockMsgWithTransform = {
-    val hash      = randomHash
-    val validator = randomHexString(32)
-    val version   = Random.nextLong()
-    val timestamp = Random.nextLong()
-    val parents   = (0 to Random.nextInt(3)) map (_ => randomHash)
-    val justifications = (0 to Random.nextInt(10)) map (
-        _ => Justification(randomHash, randomHash)
-    )
-
-    val blockMsg = BlockMessage(blockHash = hash)
-      .withHeader(
-        Header()
-          .withParentsHashList(parents)
-          .withVersion(version)
-          .withTimestamp(timestamp)
-      )
-      .withSender(validator)
-
-    BlockMsgWithTransform(Some(blockMsg), Seq.empty)
-  }
-
-  //This is needed because the alternative `Iterator.continually(elems).flatten`
-  //is not thread-safe, so it can throw OOB exception during the execution
-  def repeatedIteratorFrom[A](elems: IndexedSeq[A]): Iterator[A] =
-    Iterator
-      .iterate((0, elems(0))) {
-        case (n, e) =>
-          val next = (n + 1) % elems.length
-          (next, elems(next))
-      }
-      .map(_._2)
 }
